Drop the double overloads of the Java and C# setters
Java doubled 74 float setters and constructors with a double overload
that only cast, C# 63, and the coverage was arbitrary: Point had seven,
Cell none. Swift and Go never had them, since their literals take the
parameter's type. Every coordinate and size is a float now, so a Java
or C# caller writes 50f or casts a double.

The methods that took only a double take a float: the Container and
Stamp rotations in both ports, and C# Cell.SetWidth, Table.SetColumnWidth
and Page.DrawLine, DrawRect and FillRect. The internal callers that
computed in double cast the result, as the removed overloads did, so
every example PDF is unchanged in the four ports. The eight example
call sites that passed a double write 7f, 1f, 2.0f or declare a float,
and the README says how the ports type coordinates.

check-examples.sh is clean.

+++ com/pdfjet/CalendarMonth.java
@@ -138,8 +138,8 @@ public class CalendarMonth implements Drawable {
                         page.drawEllipse(
                                 x1 + col*dx + dx/2,
                                 y1 + row*dy + (f2.getBodyHeight()/2),
-                                dx/2.5,
-                                dy/2.5);
+                                (float) (dx/2.5),
+                                (float) (dy/2.5));
                     }
                 }
             }
